Prevent redundant imports
