+++ src/proteus/plot/cpl_atmosphere.py
@@ -10,7 +10,7 @@ import numpy as np
 from cmcrameri import cm
 from matplotlib.ticker import LogLocator
 
-from proteus.atmos_clim.janus import read_ncdfs
+from proteus.atmos_clim.common import read_ncdfs
 from proteus.utils.plot import latex_float, sample_output
 
 if TYPE_CHECKING:
